Make changes allowing same key merging
This allows included files to update parameters in their includer's
namespace if there was a maching value.

The merging only happens if the included file has matching keys,
unmatched keys revert to default functionality.

** In the case of dictionary values for matching keys, the merge function
is recursive.
** In the case of list values for matching keys, the included values are
appended to a copy of the base values.
** In the case of single values, the original value is kept.

This has a number of limitations that need to be known.

In the following example, we see that the two values are both lists
containing dictionaries. These are appened due to no easy way to denote
otherwise.

```parent.yml

settings.people:
-
  name: Alex
  month: March
-
  name: Dave
  month: April

include: child.yml
```

```child.yml
settings.people:
-
  name: Alex
  month: October
```

Merge together to result in:
```result.yml
settings.people:
-
  name: Alex
  month: March
-
  name: Dave
  month: April
-
  name: Alex
  month: October
```

Another limitation of this is the ability of overriding single values:
```parent.yml
settings:
  name: Alex

include: child.yml
```

```child.yml
settings:
  name: Bob
```

Results in:
```result.yml
settings:
  name: Alex
```

This may change in the future, I have to decide if we want the include
file to overwrite values.

The last known limitation is the ability to unset parent values. There
is currently no syntax to overwrite the parent instead of merging. Or no
syntax to subtract instead of add.

A use case for this might be specific projects unsetting an app for a
specific engine.